drop StaticBase (avoid preliminary optimization)
